requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
dj-database-url==0.5.0
Django==3.2.9
django-admin-interface==0.18.6
django-auditLog==1.0.0
django-Basic-CRUD-Generator==0.2
django-coLorfieLd==0.6.3
django-extensions==3.1.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jsonfield-backport==1.0.4
gunicorn==20.1.0
idna==3.0
pillow==9.0.0
psycopg2==2.9.3
pydotplus==2.0.2
pyparsing==3.0.6
python-dateutil>=2.6.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
djangorestframework~=3.13.1